+++ mirage/views/utils/index.ts
@@ -43,6 +43,9 @@ export function filter(model: ModelInstance, request: Request) {
                     if (field === 'id') {
                         return compareIds(model.id, val, toOperator(operator));
                     }
+                    if (model[`${field}Ids`] || model[`${field}Id`]) {
+                        return compareIds(model[`${field}Ids`] || model[`${field}Id`], val, toOperator(operator));
+                    }
                     return compare(model[field], val, toOperator(operator));
                 });
             }