+++ CHANGELOG.md
@@ -5,9 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Components
     - `schema-chunk` - yields different types of schema-chunk pieces
+    - `hierarchical-list`
+        - `hierarchical-list/item`
+        - `hieararchical-list/item-manager`
+    - `partial-registration-modal` which implements `hierarchical-list`
+        - `partial-registration-modal/manager`
+- Tests
+    - Integration
+        - `hierarchical-list`
+        - `partial-registration-modal`
 
 ### Changed
 - Components
@@ -29,6 +39,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `searchCollections`
         - added ability to filter by collection metadata
         - added ability to sort collected items by dateModified
+    - `mirage/factories/node.ts`
+        - modified the factory to set `root` to self by default
+    - `mirage/views/utils/index.ts`
+        - modified filter funtion to filter by model id
 - Packages
     - update to [ember-angle-bracket-invocation-polyfill@^2.0.2](https://github.com/rwjblue/ember-angle-bracket-invocation-polyfill/releases/tag/v2.0.2)